requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 任务队列
celery==5.3.4
redis==5.0.1

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# 配置和工具
pyyaml==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP客户端
requests==2.31.0
aiohttp==3.9.1

# 图像处理
pillow==10.1.0

# 日志和监控
structlog==23.2.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1

# 系统性能监控
psutil==5.9.5
